refactor(checker): split get_type_of_identifier megafunction into named phases (#13418)
Decompose the ~1480-line resolved-symbol body of
get_type_of_identifier_with_request out of identifier/core.rs into a new
identifier/resolved.rs with a thin orchestrator (type_of_resolved_value_symbol)
and four named phase helpers (pre-flag meaning, symbol-flag meaning guards,
merged value-symbol resolution, flow/result), plus three small helpers in
core.rs (check_identifier_strict_mode_reserved, arguments_identifier_type,
type_parameter_identifier_value).

Behavior-preserving: phase bodies are verbatim moves; the orchestrator computes
the symbol-flag tuple once and calls the phases in the original order. core.rs
drops from 1999 to 535 lines and resolved.rs is 1601 lines, both well under the
2000-line cap, leaving headroom for future identifier-resolution work.
